% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{addFoldChange}
\alias{addFoldChange}
\title{Add fold change assuming no log transform upstream}
\usage{
addFoldChange(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Add fold change assuming no log transform upstream
}
