% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{addMappedCondition}
\alias{addMappedCondition}
\title{Convert ImmuneSpace conditon-studied to a curated version}
\usage{
addMappedCondition(pd)
}
\arguments{
\item{pd}{phenotypic meta-data data.table with condtion from ImmuneSpace}
}
\description{
Convert ImmuneSpace conditon-studied to a curated version
}
