% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{addStudy}
\alias{addStudy}
\title{Extract study id from participant id and add as new field}
\usage{
addStudy(pd)
}
\arguments{
\item{pd}{phenotypic meta-data data.table with participant_id from ImmuneSpace}
}
\description{
Extract study id from participant id and add as new field
}
