% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_data}
\alias{check_data}
\title{Check data}
\usage{
check_data(data_name, data)
}
\arguments{
\item{data_name}{a character. data name to process: blood_transcript_modules,
gene_expression, gene_signatures}

\item{data}{a data.frame.}
}
\value{
A logical.
}
\description{
Check data
}
\examples{
\dontrun{
check_data("gene_expression", data)
}
}
