% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_table}
\alias{check_table}
\title{Check data}
\usage{
check_table(data_name, data)
}
\arguments{
\item{data_name}{a character. data name to process: blood_transcript_modules,
gene_signatures, or gene_expression}

\item{data}{a data.table.}
}
\value{
A logical.
}
\description{
Check data
}
\examples{
\dontrun{
check_table("blood_transcription_modules", data)
check_table("gene_signatures", data)
check_table("gene_expression", data)
}
}
