% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{correctTimeUnits}
\alias{correctTimeUnits}
\title{Convert 'Hours' or 'Months' based study times to 'Days'}
\usage{
correctTimeUnits(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Convert 'Hours' or 'Months' based study times to 'Days'
}
