% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{create_summary_data}
\alias{create_summary_data}
\title{Create a summary data object for easy use with app}
\usage{
create_summary_data(eset, analyte_type)
}
\arguments{
\item{eset}{eset object with cohort column and gene or btm expression}
}
\description{
Create a summary data object for easy use with app
}
