% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{createSummaryData}
\alias{createSummaryData}
\title{Create a summary data object for easy use with app}
\usage{
createSummaryData(eset, logData)
}
\arguments{
\item{eset}{expressionSet object with cohort column and gene or btm expression}

\item{logData}{boolean is the data in log2 format}
}
\description{
Create a summary data object for easy use with app
}
