% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{map_condition}
\alias{map_condition}
\title{Convert ImmuneSpace condition-studied to a curated version}
\usage{
map_condition(pd)
}
\arguments{
\item{pd}{phenotypic meta-data data.table with condition from ImmuneSpace}
}
\description{
Convert ImmuneSpace condition-studied to a curated version
}
