% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{matchBaselineAndPostTimepoints}
\alias{matchBaselineAndPostTimepoints}
\title{Create dt with subject * analyte matched columns for baseline and max(post) values}
\usage{
matchBaselineAndPostTimepoints(dt, analyteCol, valueCol)
}
\arguments{
\item{dt}{meta-data data.table}

\item{analyteCol}{analyte column name}

\item{valueCol}{value column name}
}
\description{
Create dt with subject * analyte matched columns for baseline and max(post) values
}
