% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{process_cohorts}
\alias{process_cohorts}
\title{Create a metadata table for cohorts}
\usage{
process_cohorts()
}
\value{
data.table
}
\description{
Create a metadata table for cohorts
}
\examples{
\dontrun{
cohorts <- process_data("cohorts")
validate(cohorts)
res <- update_table(cohorts)
}
}
