% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(data_name)
}
\arguments{
\item{data_name}{a character. data name to process: blood_transcript_modules,
gene_signatures, or gene_expression_summaries}
}
\value{
A data.table.
}
\description{
Process data
}
\examples{
\dontrun{
data <- process_data("gene_expression_summaries")
}
}
