% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{process_gene_expression}
\alias{process_gene_expression}
\title{Process gene expression data summarized by gene, blood transcription module,
and gene signature.}
\usage{
process_gene_expression()
}
\value{
data.table
}
\description{
Process gene expression data summarized by gene, blood transcription module,
and gene signature.
}
\examples{
\dontrun{
ge <- process_gene_expression()
check_table(ge)
res <- update_table("gene_expression", ge)
}
}
