% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_summaries.R
\name{process_gene_expression_summaries}
\alias{process_gene_expression_summaries}
\title{Process gene expression data summarized by gene, blood transcription module,
and gene signature.}
\usage{
process_gene_expression_summaries()
}
\value{
data.table
}
\description{
Process gene expression data summarized by gene, blood transcription module,
and gene signature.
}
\examples{
\dontrun{
summaries <- process_data("gene_expression_summaries")
validate(summaries)
res <- update_table(summaries)
}
}
