% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{update_table}
\alias{update_table}
\title{Update table}
\usage{
update_table(data_name, data)
}
\arguments{
\item{data_name}{a character. data name to process: blood_transcript_modules,
gene_expression, gene_signatures}

\item{data}{a data.frame.}
}
\value{
A list.
}
\description{
Update table
}
\examples{
\dontrun{
res <- update_table("gene_expression", data)
}
}
