% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{validate}
\alias{validate}
\title{Validate data}
\usage{
validate(data, schema_name = "analyte_explorer")
}
\arguments{
\item{data}{a data.table.}

\item{schema_name}{a character.}
}
\value{
A logical.
}
\description{
Validate data
}
\examples{
\dontrun{
validate(data)
}
}
