% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signatures.R
\name{process_gene_signatures}
\alias{process_gene_signatures}
\title{Create a metadata table for gene signatures}
\usage{
process_gene_signatures()
}
\value{
data.table
}
\description{
Updates the gene symbols with Hugo, maps pathogen to the disease types
studied, and standardizes timepoint unit to days in gene signatures table
from HIPC Dashboard. This table is used to filter the gene signatures and
display metadata about the selected gene signature (or gene) in the app.
}
\examples{
\dontrun{
library(UpdateAnno)
gs <- process_gene_signatures()
check_table(gs)
res <- update_table("gene_signatures", gs)
}
}
\references{
https://github.com/floratos-lab/hipc-dashboard-pipeline/tree/master/reformatted_data
}
